requirements.txt
asgiref==3.8.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
cffi==1.17.1
constantly==23.10.4
cryptography==43.0.1
daphne==4.1.2
Django==5.1
django-rest-framework==0.1.0
djangorestframework==3.15.2
hyperlink==21.0.0
icalendar==5.0.13
idna==3.8
incremental==24.7.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pyOpenSSL==24.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
service-identity==24.1.0
setuptools==74.1.2
six==1.16.0
sqlparse==0.5.1
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.2
zope.interface==7.0.3
